+++ resource-server/src/main/java/com/inhabas/api/controller/LoginController.java
@@ -1,7 +1,7 @@
 package com.inhabas.api.controller;
 
 import com.inhabas.api.auth.utils.argumentResolver.Authenticated;
-import com.inhabas.api.domain.member.LoginMember;
+import com.inhabas.api.domain.member.security.LoginMember;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
-import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -26,7 +25,7 @@ public class LoginController {
     @GetMapping("${authenticate.oauth2-success-handle-url}")
     @Operation(description = "로그인 성공하여 최종적으로 accessToken, refreshToken 을 발행한다.", hidden = true)
     public ResponseEntity<?> successLogin(
-            HttpServletRequest request, @Authenticated LoginMember authUserDetail) throws URISyntaxException {
+            HttpServletRequest request, @Authenticated LoginMember authUserDetail) {
 
         request.getSession().invalidate();
         throw new NotImplementedException("소셜로그인 다시 구현해야함");
