+++ resource-server/src/main/java/com/inhabas/api/controller/SignUpController.java
@@ -1,7 +1,7 @@
 package com.inhabas.api.controller;
 
 import com.inhabas.api.auth.utils.argumentResolver.Authenticated;
-import com.inhabas.api.domain.member.LoginMember;
+import com.inhabas.api.domain.member.security.LoginMember;
 import com.inhabas.api.dto.member.MajorInfoDto;
 import com.inhabas.api.dto.signUp.AnswerDto;
 import com.inhabas.api.dto.signUp.MemberDuplicationQueryCondition;
